requirements.txt
fastapi
uvicorn
sentence-transformers
torch
google-generativeai
